Refactor vectorLSM logic, migrate merge related logic
